fix: Clean up
